Allow using i2c as bus number if not None
This patch allows one to use the i2c argument when initializing the BNO055 device
as the bus number.  The default is to not pass an i2c argument, which will result
in the default i2c device to be used.  This is useful in the event that one wants
to override the default i2c device (such as in the case of using i2c-gpio on
raspberry pi).